Render null-safe badge on package page + refactoring.
