+++ app/lib/frontend/templates/views/pkg/badge.mustache
@@ -1,0 +1,10 @@
+{{! Copyright (c) 2020, the Dart project authors.  Please see the AUTHORS file
+    for details. All rights reserved. Use of this source code is governed by a
+    BSD-style license that can be found in the LICENSE file. }}
+
+<span class="package-badge"{{#title}} title="{{ title }}"{{/title}}>
+  {{#icon}}
+  <img class="package-badge-icon" src="{{ icon }}" />
+  {{/icon}}
+  {{ label }}
+</span>
